requirements.txt
torch==2.8.0
transformers==4.43.3
accelerate>=1.10,<2
gradio>=4.30,<5
python-dotenv>=1.0.1,<2
pandas>=2.2,<3
openpyxl>=3.1.5,<4
requests>=2.32.3,<3
urllib3[socks]>=2.5.0,<3
beautifulsoup4>=4.12,<5
lxml>=5.2,<7
fake-useragent>=2.2,<3
tenacity>=8.2,<9
tiktoken==0.11.0
bcrypt>=4.1,<5
psutil>=7,<8
googlemaps>=4.10,<5
googlesearch-python>=1.2,<2
xlrd==2.0.1
google-search-results>=2.4,<3
huggingface_hub>=0.20,<1

# Backend dependencies
fastapi>=0.104,<1
uvicorn[standard]>=0.24,<1
sqlalchemy>=2.0,<3
pydantic>=2.5,<3
python-multipart>=0.0.6
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
pyjwt>=2.8,<3
email-validator>=2.1,<3
